- Split Debug classes in separate files - Free Camera controls are now disabled when debug menu is enabled.
