+++ ScriptableRenderPipeline/HDRenderPipeline/Debug/LightingDebug.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f16fb47b338669345b3a0d4156d8f900
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
