+++ ScriptableRenderPipeline/HDRenderPipeline/Debug/MaterialDebug.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 376e999a68d3f5843949eb37e24c72f0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
