+++ ScriptableRenderPipeline/HDRenderPipeline/Debug/RenderingDebug.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1e64277580777ed4fae930253871a755
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
